requirements.txt
anthropic>=0.25.0
python-dotenv>=1.0.0
requests>=2.31.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0
pymupdf>=1.24.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
